refactor: 인풋, 아웃풋으로 뷰 분리
